requirements.txt
pytz
Pillow
markdown
boto
psycopg2
Django
django-registration
django-taggit
django-taggit-autosuggest
django-contrib-comments
django-vote
social-auth-app-django
markdown2==2.3.1
django-markdown-deux
django-pagedown
djangorestframework
django-filter
django-gravatar2
django-versionfield2
django-modeltranslation
django-storages
django-notifications-hq
